# swagger.yml
openapi: 3.0.0
info:
  title: Trail Management API
  version: 1.0.0
  description: API for managing trails, locations, trail views, and users.
servers:
  - url: "/api"
    description: Local development server
paths:
...
  /users:
    get:
      summary: Get all users
      operationId: "user.get_all_users"
      tags:
        - Users
      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Create a new user
      operationId: "user.create_user"
      tags:
        - Users
      requestBody:
        description: User data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/login:
    post:
      summary: User login
      operationId: "user.login_user"
      tags:
       - Users
      requestBody:
        description: User login credentials
        required: true
        content:
          application/json:
           schema:
             $ref: '#/components/schemas/UserLogin'
      responses:
        200:
          description: User login successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLogin'
  /users/update/{user_id}:
    put:
      summary: Update a user by ID
      operationId: "user.update_user"
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Updated user data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    delete:
      summary: Delete a user by ID
      operationId: "user.delete_user"
      tags:
        - Users
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: User deleted successfully
openapi: 3.0.0
info:
  title: Example API
  version: 1.0.0
components:
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
          description: Unique ID for the user.
          maxLength: 8
        user_name:
          type: string
          description: Name of the user.
          maxLength: 225
        email:
          type: string
          description: Email of the user.
          maxLength: 225
        password:
          type: string
          description: Password of the user.
          maxLength: 225
        phone_number:
          type: integer
          description: Phone number of the user.
      required:
        - user_id
        - email
        - password
      example:
        user_id: "USER1234"
        user_name: "Alice Johnson"
        email: "alice@example.com"
        password: "MyPass"
        phone_number: 1234567890
    UserUpdate:
      type: object
      properties:
        user_name:
          type: string
          description: Updated name of the user.
          maxLength: 225
        email:
          type: string
          description: Updated email of the user.
          maxLength: 225
        phone_number:
          type: integer
          description: Updated phone number of the user.
      example:
        user_name: "Alice J."
        email: "alice.j@example.com"
        phone_number: 9876543210
    UserLogin:
      type: object
      properties:
        user_name:
          type: string
          description: Login Credentials.
          maxLength: 225
        email:
          type: string
          description: email of the user.
          maxLength: 225
        password:
          type: string
          description: Password of the user.
      example:
        password: "Alice J"
        email: "alice.j@example.com"
        

    # UserLogin:
    #   type: object
    #   properties:
    #     email:
    #       type: string
    #       format: email
    #       example: user@example.com
    #     password:
    #       type: string
    #       format: password
    #       example: password123